Made changes per PR comments
Modified schedulerActive to scale well

Moved policies text to seperate folder in static

Added CSM link blue and used variables

converted site-subtitle-link from ID to class in style.css and App.tsx

Fixed policies link not coloring when on page
